api_search/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
nltk>=3.8.0
pydantic>=2.0.0
